llm_requirements.txt

# Requerimientos para LLM, análisis financiero y extracción de noticias

# LLM y NLP
transformers>=4.20.0
torch>=1.12.0
openai>=1.0.0           # OpenAI GPT
anthropic>=0.3.0         # Claude
huggingface-hub>=0.15.0  # HuggingFace
ollama                   # Modelos locales (opcional)

# Extracción de noticias y datos financieros
tweepy>=4.0.0
feedparser>=6.0.0
pygooglenews>=0.1.2
yfinance>=0.2.0
requests>=2.28.0
python-dateutil>=2.8.0

# Análisis y utilidades
pandas>=1.4.0
numpy>=1.21.0

# Instalación recomendada:
# pip install -r llm_requirements.txt
# Para Ollama local:
# pip install ollama
# o
# curl -fsSL https://ollama.ai/install.sh | sh
